Polish agent docs surfaces to a clean afdocs 100.
Keep crypto out of Edge middleware, drop XML/txt non-pages from the
sampled set, match homepage markdown to HTML segments, and give static
HTML pages an hourly revalidate so cache-header-hygiene passes.
