+++ website/app/api/page-markdown/[...path]/route.ts
@@ -2,7 +2,7 @@ import {
   renderHomeMarkdown,
   renderProductMarkdown,
 } from '../../../lib/page-markdown'
-import { textResponse } from '../../../lib/agent-docs'
+import { textResponse } from '../../../lib/text-response'
 import { isProductPageId } from '../../../lib/product-pages'
 import type { AppLocale } from '../../../i18n/resources'
