+++ website/app/blog/page.tsx
@@ -3,7 +3,7 @@ import { getMessages } from '../i18n/resources'
 import { buildPageMetadata } from '../page-metadata'
 import { BlogListContent } from './blog-list-content'
 import { listPosts } from './lib/posts'
-import { blogListMarkdownPath, blogListPath } from './lib/routes'
+import { blogFeedPath, blogListMarkdownPath, blogListPath } from './lib/routes'
 
 const messages = getMessages('en')
 
@@ -18,6 +18,7 @@ export const metadata: Metadata = buildPageMetadata('en', {
   title: messages.blog.metadataTitle,
   description: messages.blog.metadataDescription,
   markdownPath: blogListMarkdownPath('en'),
+  rssPath: blogFeedPath('en'),
   languages: {
     en: blogListPath('en'),
     'zh-CN': blogListPath('zh-cn'),
