+++ website/app/install/page.tsx
@@ -1,6 +1,8 @@
 import type { Metadata } from 'next'
 import { productPageMetadata, renderProductPage } from '../lib/product-route'
 
+export const revalidate = 3600
+
 export const metadata: Metadata = productPageMetadata('en', 'install')
 
 export default function InstallPage() {
