+++ website/app/llms.txt/route.ts
@@ -1,5 +1,5 @@
 import { renderLlmsTxt } from '../llms-txt'
-import { textResponse } from '../lib/agent-docs'
+import { textResponse } from '../lib/text-response'
 
 export const revalidate = 300
