+++ website/app/page.tsx
@@ -3,6 +3,8 @@ import { HomePageContent } from './home-page-content'
 import { getMessages } from './i18n/resources'
 import { buildPageMetadata } from './page-metadata'
 
+export const revalidate = 3600
+
 export const metadata: Metadata = buildPageMetadata('en', { pathname: '/' })
 
 export default function HomePage() {
