+++ website/app/sitemap.ts
@@ -1,14 +1,13 @@
 import type { MetadataRoute } from 'next'
-import { siteConfig } from './site-config'
-import {
-  absoluteUrl,
-  feedSitemapPaths,
-  listPublicPages,
-} from './lib/public-pages'
+import { absoluteUrl, listPublicPages } from './lib/public-pages'
 
 /*
   Not force-static: the sitemap has to pick up posts that publish themselves
   between deploys, so it re-renders on the same window as the blog.
+
+  Feeds stay out of the sitemap: afdocs treats every sitemap URL as a "page",
+  and an XML feed cannot carry the HTML llms.txt directive. Blog indexes
+  advertise RSS via rel=alternate instead.
 */
 export const revalidate = 3600
 
@@ -16,7 +15,7 @@ export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
   const lastModified = new Date()
   const pages = await listPublicPages()
 
-  const entries: MetadataRoute.Sitemap = pages.flatMap((page) => [
+  return pages.flatMap((page) => [
     {
       url: absoluteUrl(page.htmlPath),
       lastModified: page.lastModified ?? lastModified,
@@ -30,21 +29,4 @@ export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
       priority: Math.max(0.3, page.priority - 0.3),
     },
   ])
-
-  entries.push(
-    {
-      url: `${siteConfig.url}${siteConfig.llmsFullTxtPath}`,
-      lastModified,
-      changeFrequency: 'weekly',
-      priority: 0.4,
-    },
-    ...feedSitemapPaths().map((path) => ({
-      url: absoluteUrl(path),
-      lastModified,
-      changeFrequency: 'weekly' as const,
-      priority: 0.3,
-    })),
-  )
-
-  return entries
 }
