+++ website/app/zh-cn/page.tsx
@@ -3,6 +3,8 @@ import { HomePageContent } from '../home-page-content'
 import { getMessages } from '../i18n/resources'
 import { buildPageMetadata } from '../page-metadata'
 
+export const revalidate = 3600
+
 export const metadata: Metadata = buildPageMetadata('zh-cn', { pathname: '/zh-cn' })
 
 export default function ChineseHomePage() {
